+++ src/test/java/dk/nsi/sdm4/vitamin/parser/VitaminParserSinglefilesIntegrationTest.java
@@ -29,6 +29,7 @@ package dk.nsi.sdm4.vitamin.parser;
 import dk.nsi.sdm4.core.persistence.recordpersister.RecordPersister;
 import dk.nsi.sdm4.core.persistence.recordpersister.RecordSpecification;
 import dk.nsi.sdm4.testutils.TestDbConfiguration;
+import dk.nsi.sdm4.vitamin.VitaminimporterApplicationTestConfig;
 import dk.nsi.sdm4.vitamin.config.VitaminimporterApplicationConfig;
 import dk.nsi.sdm4.vitamin.recordspecs.VitaminRecordSpecs;
 import org.joda.time.DateTime;
@@ -54,7 +55,7 @@ import static org.junit.Assert.fail;
  */
 @RunWith(SpringJUnit4ClassRunner.class)
 @Transactional
-@ContextConfiguration(classes = {VitaminimporterApplicationConfig.class, TestDbConfiguration.class})
+@ContextConfiguration(classes = {VitaminimporterApplicationTestConfig.class, TestDbConfiguration.class})
 public class VitaminParserSinglefilesIntegrationTest
 {
 	@Autowired